Change main screen

+++ app/screens/createwallet/CreateWalletByEmailScreen.js
@@ -6,6 +6,7 @@ import {
   TouchableWithoutFeedback,
   AsyncStorage,
   Image,
+  ScrollView,
 } from 'react-native';
 import { CheckBox } from 'react-native-elements';
 import Toast from 'react-native-root-toast';
@@ -37,7 +38,12 @@ export default class CreateWalletByEmailScreen extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      isChecked: true,
+      isChecked: false,
+      createWalletInfo: {
+        email: null,
+        password: null,
+        passwordConfirm: null,
+      },
     };
   }
 
@@ -46,11 +52,6 @@ export default class CreateWalletByEmailScreen extends Component {
 
     this.setState({
       isChecked: !isChecked,
-      createWalletInfo: {
-        email: null,
-        password: null,
-        passwordConfirm: null,
-      },
     });
   }
 
@@ -63,21 +64,40 @@ export default class CreateWalletByEmailScreen extends Component {
     });
   }
 
+  validateEmail = (email) => {
+    const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+    return re.test(String(email).toLowerCase());
+  }
+
   _handleClickCreateWallet = async () => {
     const { navigation } = this.props;
+    const { isChecked, createWalletInfo } = this.state;
+
     try {
+      if (!isChecked) {
+        throw new Error('Please read and accept terms and conditions!');
+      }
+
+      if (!this.validateEmail(createWalletInfo.email)) {
+        throw new Error('Email is not valid!');
+      }
+      if (!createWalletInfo.password || (createWalletInfo.password !== createWalletInfo.passwordConfirm)) {
+        throw new Error('Password must match password confirmation!');
+      }
+
       const generatedMnemonic = bip39.generateMnemonic();
       const seed = bip39.mnemonicToSeedHex(generatedMnemonic);
       const generatedPrivateKey = hdkey.fromMasterSeed(seed).privateKey.toString('hex');
 
       const {
         privateKey, address, mnemonic, keystore,
-      } = await WalletService.importWalletFromPrivateKey('nanj', generatedPrivateKey, generatedMnemonic, '123456');
+      } = await WalletService.importWalletFromPrivateKey('nanj', generatedPrivateKey, generatedMnemonic, createWalletInfo.password);
 
-      AsyncStorage.setItem('address', address);
-      AsyncStorage.setItem('privateKey', privateKey);
-      AsyncStorage.setItem('mnemonic', mnemonic);
-      AsyncStorage.setItem('keystore', keystore);
+      await AsyncStorage.setItem('address', address);
+      await AsyncStorage.setItem('privateKey', privateKey);
+      await AsyncStorage.setItem('mnemonic', mnemonic);
+      await AsyncStorage.setItem('keystore', keystore);
+      console.log('keystore', keystore);
 
       navigation.navigate('MainScreen');
     } catch (error) {
@@ -144,9 +164,10 @@ export default class CreateWalletByEmailScreen extends Component {
     const { isChecked } = this.state;
 
     return (
-      <View style={styles.checkBoxAccept}>
+      <View style={styles.termAndConditionContainer}>
         <CheckBox
-          containerStyle={styles.checkboxs}
+          containerStyle={styles.checkboxContainer}
+          iconRight
           checked={isChecked}
           checkedIcon="check-square"
           uncheckedIcon="square"
@@ -154,13 +175,12 @@ export default class CreateWalletByEmailScreen extends Component {
           onPress={() => this._handleToggleCheckBox()}
         />
         <TouchableWithoutFeedback
-          activeOpacity={0.5}
           onPress={() => this._handleToggleCheckBox()}
         >
           <View
-            style={styles.touchIAccept}
+            style={styles.textTermAndConditions}
           >
-            <Text style={styles.textIAccept}>{I18n.t('createByPhoneNumber.iAccept')}</Text>
+            <Text style={styles.textAccept}>{I18n.t('createByPhoneNumber.iAccept')}</Text>
             <Text style={styles.textTerms}>{I18n.t('createByPhoneNumber.termsAndConditions')}</Text>
           </View>
         </TouchableWithoutFeedback>
@@ -171,7 +191,7 @@ export default class CreateWalletByEmailScreen extends Component {
   _renderButtonCreate = () => (
     <MangoGradientButton
       btnText={I18n.t('signin.title')}
-      btnStyle={styles.btnSigninContainer}
+      btnStyle={styles.btnCreateWalletContainer}
       onPress={() => this._handleClickCreateWallet()}
     />
   )
@@ -179,8 +199,10 @@ export default class CreateWalletByEmailScreen extends Component {
   render() {
     return (
       <View style={styles.container}>
-        {this._renderFormCreateByEmail()}
-        {this._renderTermsAndConditions()}
+        <ScrollView>
+          {this._renderFormCreateByEmail()}
+          {this._renderTermsAndConditions()}
+        </ScrollView>
         {this._renderButtonCreate()}
       </View>
     );
@@ -195,34 +217,31 @@ const styles = ScaledSheet.create({
     backgroundColor: '#F5F7FA',
   },
 
-  checkboxs: {
-    alignItems: 'center',
-    backgroundColor: '#F5F7FA',
-    borderWidth: 0,
-    width: '50@s',
+  termAndConditionContainer: {
+    flexDirection: 'row',
+    marginTop: '10@s',
   },
 
-  textIAccept: {
-    color: '#4d4e4e',
-    fontSize: '16@s',
+  checkboxContainer: {
+    backgroundColor: 'transparent',
+    padding: '0@s',
+    borderWidth: 0,
+  },
+
+  textTermAndConditions: {
+    flexDirection: 'row',
     alignItems: 'center',
     alignSelf: 'center',
   },
 
-  checkBoxAccept: {
-    flexDirection: 'row',
-    marginTop: '15@s',
+  textAccept: {
+    color: '#121313',
+    fontSize: '16@ms',
   },
 
   textTerms: {
-    color: '#6580ce',
-    fontSize: '16@s',
-    alignItems: 'center',
-    alignSelf: 'center',
-  },
-
-  touchIAccept: {
-    flexDirection: 'row',
+    color: '#1f45b9',
+    fontSize: '16@ms',
   },
 
   formLoginContainer: {
@@ -266,7 +285,8 @@ const styles = ScaledSheet.create({
     fontWeight: '100',
   },
 
-  btnSigninContainer: {
+  btnCreateWalletContainer: {
     width: '220@s',
+    marginBottom: '24@s',
   },
 });
